create grapql using and cruede

+++ src/main/java/org/example/dto/Products.java
@@ -12,6 +12,8 @@ public class Products{
     private Integer id;
     private String name;
     private String adress;
-    private Float price;
+    private Double price;
     private String category;
+
+
 }
